Move swap from 'channel' package to 'payment' package
