+++ node/engine/store/durablestore.go
@@ -141,7 +141,7 @@ func (ds *DurableStore) GetChannelSecretKey() *[]byte {
 	return &val
 }
 
-func (ds *DurableStore) GetSwapById(id types.Destination) (channel.Swap, error) {
+func (ds *DurableStore) GetSwapById(id types.Destination) (payments.Swap, error) {
 	var sJSON string
 	err := ds.swaps.View(func(tx *buntdb.Tx) error {
 		var err error
@@ -150,18 +150,18 @@ func (ds *DurableStore) GetSwapById(id types.Destination) (channel.Swap, error) 
 	})
 
 	if errors.Is(err, buntdb.ErrNotFound) {
-		return channel.Swap{}, ErrNoSuchSwap
+		return payments.Swap{}, ErrNoSuchSwap
 	}
-	var swap channel.Swap
+	var swap payments.Swap
 	err = json.Unmarshal([]byte(sJSON), &swap)
 	if err != nil {
-		return channel.Swap{}, fmt.Errorf("error unmarshaling swap %s", id)
+		return payments.Swap{}, fmt.Errorf("error unmarshaling swap %s", id)
 	}
 
 	return swap, nil
 }
 
-func (ds *DurableStore) SetSwap(swap channel.Swap) error {
+func (ds *DurableStore) SetSwap(swap payments.Swap) error {
 	sJSON, err := json.Marshal(swap)
 	if err != nil {
 		return err
@@ -235,7 +235,7 @@ func (ds *DurableStore) SetObjective(obj protocols.Objective) error {
 				return fmt.Errorf("error setting swap channel %s from objective %s: %w", related.Id, obj.Id(), err)
 			}
 
-		case *channel.Swap:
+		case *payments.Swap:
 			err := ds.SetSwap(*related)
 			if err != nil {
 				return fmt.Errorf("error setting swap %s from objective %s: %w", related.Id, obj.Id(), err)
@@ -888,7 +888,7 @@ func (ds *DurableStore) DestroyObjective(id protocols.ObjectiveId) error {
 	})
 }
 
-func (ds *DurableStore) GetPendingSwapByChannelId(id types.Destination) (*channel.Swap, error) {
+func (ds *DurableStore) GetPendingSwapByChannelId(id types.Destination) (*payments.Swap, error) {
 	var pendingSwapId types.Destination
 	err := ds.objectives.View(func(tx *buntdb.Tx) error {
 		err := tx.Ascend("", func(key, objJSON string) bool {
@@ -924,8 +924,8 @@ func (ds *DurableStore) GetPendingSwapByChannelId(id types.Destination) (*channe
 	return nil, nil
 }
 
-func (ds *DurableStore) GetSwapsByChannelId(id types.Destination) ([]channel.Swap, error) {
-	swapQueue := channel.NewSwapsQueue()
+func (ds *DurableStore) GetSwapsByChannelId(id types.Destination) ([]payments.Swap, error) {
+	swapQueue := payments.NewSwapsQueue()
 
 	err := ds.channelToSwaps.View(func(tx *buntdb.Tx) error {
 		sJSON, err := tx.Get(id.String())
@@ -945,7 +945,7 @@ func (ds *DurableStore) GetSwapsByChannelId(id types.Destination) ([]channel.Swa
 		return nil, err
 	}
 
-	var swapsToReturn []channel.Swap
+	var swapsToReturn []payments.Swap
 	swaps := swapQueue.Values()
 	for _, swap := range swaps {
 		s, err := ds.GetSwapById(swap.Id)
@@ -958,8 +958,8 @@ func (ds *DurableStore) GetSwapsByChannelId(id types.Destination) ([]channel.Swa
 	return swapsToReturn, nil
 }
 
-func (ds *DurableStore) SetChannelToSwaps(swap channel.Swap) (channel.Swap, error) {
-	swapQueue := channel.NewSwapsQueue()
+func (ds *DurableStore) SetChannelToSwaps(swap payments.Swap) (payments.Swap, error) {
+	swapQueue := payments.NewSwapsQueue()
 
 	err := ds.channelToSwaps.View(func(tx *buntdb.Tx) error {
 		sJSON, err := tx.Get(swap.ChannelId.String())
@@ -976,14 +976,14 @@ func (ds *DurableStore) SetChannelToSwaps(swap channel.Swap) (channel.Swap, erro
 		return err
 	})
 	if err != nil {
-		return channel.Swap{}, err
+		return payments.Swap{}, err
 	}
 
 	removedSwap := swapQueue.Enqueue(swap)
 
 	swapsJson, err := swapQueue.MarshalJSON()
 	if err != nil {
-		return channel.Swap{}, fmt.Errorf("error marshalling swap queue %w", err)
+		return payments.Swap{}, fmt.Errorf("error marshalling swap queue %w", err)
 	}
 
 	err = ds.channelToSwaps.Update(func(tx *buntdb.Tx) error {
@@ -991,7 +991,7 @@ func (ds *DurableStore) SetChannelToSwaps(swap channel.Swap) (channel.Swap, erro
 		return err
 	})
 	if err != nil {
-		return channel.Swap{}, err
+		return payments.Swap{}, err
 	}
 
 	return removedSwap, nil
